fix(ios): Swift filename ve type cakismalarini gider [agent:developer]
- ProjectMapper.swift -> ProjectStatusMapper.swift (Home ile cakisma)
- Project.swift silindi (Home/Domain/Models/Project.swift ile birlesti)
- ProjectRepositoryProtocol -> ProjectStatusRepositoryProtocol
- Home/Project modeli superset olarak guncellendi (pending, techStack vb.)
- ProjectStatus'a archived eklendi
- Tum switch statement'lara archived case eklendi

Refs: #32
